requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==9.0.0
pydantic==2.9.2
pytest~=8.3.5
h11~=0.14.0
pip~=23.2.1
typing_extensions~=4.13.2
sniffio~=1.3.1
starlette~=0.38.6
pydantic_core~=2.23.4
requests~=2.32.3
anyio~=4.9.0
click~=8.1.8
python-multipart~=0.0.20
httpx~=0.28.1